+++ clients/fides-js/src/fides.ts
@@ -210,7 +210,6 @@ _Fides = {
     preventDismissal: false,
     allowHTMLDescription: null,
     base64Cookie: false,
-    fidesPreviewMode: false,
     fidesPrimaryColor: null,
   },
   fides_meta: {},